+++ src/main/java/com/aswemake/market/config/SecurityConfig.java
@@ -1,0 +1,45 @@
+package com.aswemake.market.config;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
+import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.SecurityFilterChain;
+
+@Configuration
+@EnableWebSecurity
+@EnableMethodSecurity
+@RequiredArgsConstructor
+public class SecurityConfig {
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
+        http
+                .csrf(AbstractHttpConfigurer::disable)
+                .authorizeHttpRequests(request -> request
+                        .requestMatchers("/api/members/sign-up").anonymous()
+                        .anyRequest().authenticated()
+                )
+        ;
+        http
+                .formLogin(form -> form
+                        .loginProcessingUrl("/api/members/sign-in")
+                        .usernameParameter("userId")
+                        .defaultSuccessUrl("/api/members/test") //TODO success / failure handler 추가
+                        .failureUrl("/api/members/sign-in-fail")
+                )
+        ;
+
+        return http.build();
+    }
+}
