+++ src/main/java/com/aswemake/market/exception/ApplicationException.java
@@ -1,0 +1,17 @@
+package com.aswemake.market.exception;
+
+import lombok.Getter;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class ApplicationException extends RuntimeException {
+
+    private final HttpStatus httpStatus;
+    private final ErrorCode errorCode;
+
+    public ApplicationException(HttpStatus httpStatus, ErrorCode errorCode, String message) {
+        super(message);
+        this.errorCode = errorCode;
+        this.httpStatus = httpStatus;
+    }
+}
