+++ src/main/java/com/aswemake/market/exception/ErrorCode.java
@@ -1,0 +1,6 @@
+package com.aswemake.market.exception;
+
+public enum ErrorCode {
+    INVALID_REQUEST,
+    DUPLICATE_USERID_ERROR
+}
