+++ src/main/java/com/aswemake/market/exception/ErrorResponseDto.java
@@ -1,0 +1,12 @@
+package com.aswemake.market.exception;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public class ErrorResponseDto {
+
+    private ErrorCode errorCode;
+    private String message;
+}
