+++ src/main/java/com/aswemake/market/member/application/security/CustomUserDetailsService.java
@@ -1,0 +1,26 @@
+package com.aswemake.market.member.application.security;
+
+import com.aswemake.market.member.domain.entity.Member;
+import com.aswemake.market.member.domain.repository.MemberRepository;
+import com.aswemake.market.member.domain.vo.UserId;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
+
+@Service("userDetailsService")
+@RequiredArgsConstructor
+public class CustomUserDetailsService implements UserDetailsService {
+
+    private final MemberRepository memberRepository;
+
+    @Override
+    public UserDetails loadUserByUsername(String userId) throws UsernameNotFoundException {
+
+        Member member = memberRepository.findByUserId(UserId.of(userId))
+                .orElseThrow(() -> new UsernameNotFoundException("사용자 정보를 다시 확인해주세요."));
+
+        return new AccountContext(member, member.createRole());
+    }
+}
