+++ src/main/java/com/aswemake/market/member/domain/vo/Address.java
@@ -1,0 +1,29 @@
+package com.aswemake.market.member.domain.vo;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
+public class Address {
+
+    private final String address;
+    private final String addressDetail;
+
+    private Address(String address, String addressDetail) {
+        this.address = address;
+        this.addressDetail = addressDetail;
+    }
+
+    public static Address of(String address, String addressDetail) {
+
+        if (address == null || address.isEmpty()) {
+            throw new IllegalArgumentException("주소를 입력하세요.");
+        }
+        if (addressDetail == null || addressDetail.isEmpty()) {
+            throw new IllegalArgumentException("상세 주소를 입력하세요.");
+        }
+        return new Address(address, addressDetail);
+    }
+}
