+++ src/main/java/com/aswemake/market/member/domain/vo/Password.java
@@ -1,0 +1,35 @@
+package com.aswemake.market.member.domain.vo;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PRIVATE, force = true)
+public class Password {
+
+    private final String password;
+
+    private Password(String password) {
+        this.password = password;
+    }
+
+    public static Password of(String password) {
+
+        if (password == null || password.isEmpty()) {
+            throw new IllegalArgumentException("비밀번호를 입력하세요.");
+        }
+        if (password.contains(" ")) {
+            throw new IllegalArgumentException("비밀번호엔 공백을 포함할 수 없습니다.");
+        }
+        return new Password(password);
+    }
+
+    public static Password encodedPassword(String password) {
+        return new Password(password);
+    }
+
+    public String password() {
+        return this.password;
+    }
+}
