+++ src/main/java/com/aswemake/market/member/infrastructure/repository/MemberQueryRepository.java
@@ -1,0 +1,8 @@
+package com.aswemake.market.member.infrastructure.repository;
+
+import com.aswemake.market.member.domain.vo.UserId;
+
+public interface MemberQueryRepository {
+
+    boolean isUserIdExist(UserId userId);
+}
