--- src/main/java/com/aswemake/market/user/exception/ErrorCode.java
@@ -1,6 +1,0 @@
-package com.aswemake.market.user.exception;
-
-public enum ErrorCode {
-    INVALID_REQUEST,
-    DUPLICATE_USERID_ERROR
-}
