--- src/main/java/com/aswemake/market/user/exception/ErrorResponseDto.java
@@ -1,12 +1,0 @@
-package com.aswemake.market.user.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class ErrorResponseDto {
-
-    private ErrorCode errorCode;
-    private String message;
-}
